+++ RELEASE_NOTES.md
@@ -3,6 +3,7 @@ Release Notes
 
 ## 1.6.2
 * Functions: Support Elastic Premium SKUs for Functions service plans.
+* Storage: Support for minimum TLS version.
 * Virtual Machine: Support for customData on osProfile Properties 
 * WebApp: Added support for PrivateEndpoints